dev-requirements.txt
--editable .
responses>=0.5.0
djangorestframework-simplejwt==4.4.0
django-allauth>=0.25.0
coveralls>=1.11.1
unittest-xml-reporting>=3.0.2